+++ biopytools/mga/__init__.py
@@ -1,0 +1,8 @@
+"""MGA共识基因组组装|MGA consensus genome assembly"""
+
+from .main import MGARunner
+from .config import MGAConfig
+
+__version__ = "1.0.0"
+
+__all__ = ["MGARunner", "MGAConfig"]